Ensure all non-test Typescript files have < 600 lines (#1436)
* mutations file code reduced

* Queries.ts file line reduced

* OrgList.tsx file line reduced

* Add TSDoc comments to enhance code documentation on mutations modules

* mutations.ts file reduced to 598 lines -done

* Queries.ts file line reduced to 574 line -done

* OrgList.tsx file reduced to 568 line -done

* Delete package-lock.json

* Restore Package-lock.json back to previous State

* pull-requests.yml file count line value set to 600 -done
