+++ src/screens/OrgList/OrgList.tsx
@@ -14,14 +14,13 @@ import OrgListCard from 'components/OrgListCard/OrgListCard';
 import SuperAdminScreen from 'components/SuperAdminScreen/SuperAdminScreen';
 import type { ChangeEvent } from 'react';
 import React, { useEffect, useState } from 'react';
-import { Col, Dropdown, Form, Row } from 'react-bootstrap';
+import { Dropdown, Form } from 'react-bootstrap';
 import Button from 'react-bootstrap/Button';
 import Modal from 'react-bootstrap/Modal';
 import { useTranslation } from 'react-i18next';
 import InfiniteScroll from 'react-infinite-scroll-component';
 import { Link } from 'react-router-dom';
 import { toast } from 'react-toastify';
-import convertToBase64 from 'utils/convertToBase64';
 import { errorHandler } from 'utils/errorHandler';
 import type {
   InterfaceOrgConnectionInfoType,
@@ -29,6 +28,7 @@ import type {
   InterfaceUserType,
 } from 'utils/interfaces';
 import styles from './OrgList.module.css';
+import OrganizationModal from './OrganizationModal';
 
 function orgList(): JSX.Element {
   const { t } = useTranslation('translation', { keyPrefix: 'orgList' });
@@ -484,171 +484,29 @@ function orgList(): JSX.Element {
           </>
         )}
         {/* Create Organization Modal */}
-        <Modal
-          show={showModal}
-          onHide={toggleModal}
-          aria-labelledby="contained-modal-title-vcenter"
-          centered
-        >
-          <Modal.Header
-            className="bg-primary"
-            closeButton
-            data-testid="modalOrganizationHeader"
-          >
-            <Modal.Title className="text-white">
-              {t('createOrganization')}
-            </Modal.Title>
-          </Modal.Header>
-          <Form onSubmitCapture={createOrg}>
-            <Modal.Body>
-              <Form.Label htmlFor="orgname">{t('name')}</Form.Label>
-              <Form.Control
-                type="name"
-                id="orgname"
-                className="mb-3"
-                placeholder={t('enterName')}
-                data-testid="modalOrganizationName"
-                autoComplete="off"
-                required
-                value={formState.name}
-                onChange={(e): void => {
-                  const inputText = e.target.value;
-                  if (inputText.length < 50) {
-                    setFormState({
-                      ...formState,
-                      name: e.target.value,
-                    });
-                  }
-                }}
-              />
-              <Form.Label htmlFor="descrip">{t('description')}</Form.Label>
-              <Form.Control
-                type="descrip"
-                id="descrip"
-                className="mb-3"
-                placeholder={t('description')}
-                autoComplete="off"
-                required
-                value={formState.descrip}
-                onChange={(e): void => {
-                  const descriptionText = e.target.value;
-                  if (descriptionText.length < 200) {
-                    setFormState({
-                      ...formState,
-                      descrip: e.target.value,
-                    });
-                  }
-                }}
-              />
-              <Form.Label htmlFor="location">{t('location')}</Form.Label>
-              <Form.Control
-                type="text"
-                id="location"
-                className="mb-3"
-                placeholder={t('location')}
-                autoComplete="off"
-                required
-                value={formState.location}
-                onChange={(e): void => {
-                  const locationText = e.target.value;
-                  if (locationText.length < 100) {
-                    setFormState({
-                      ...formState,
-                      location: e.target.value,
-                    });
-                  }
-                }}
-              />
-
-              <Row className="mb-3">
-                <Col>
-                  <Form.Label htmlFor="userRegistrationRequired">
-                    {t('userRegistrationRequired')}
-                  </Form.Label>
-                  <Form.Switch
-                    id="userRegistrationRequired"
-                    data-testid="userRegistrationRequired"
-                    type="checkbox"
-                    defaultChecked={formState.userRegistrationRequired}
-                    onChange={(): void =>
-                      setFormState({
-                        ...formState,
-                        userRegistrationRequired:
-                          !formState.userRegistrationRequired,
-                      })
-                    }
-                  />
-                </Col>
-                <Col>
-                  <Form.Label htmlFor="visibleInSearch">
-                    {t('visibleInSearch')}
-                  </Form.Label>
-                  <Form.Switch
-                    id="visibleInSearch"
-                    data-testid="visibleInSearch"
-                    type="checkbox"
-                    defaultChecked={formState.visible}
-                    onChange={(): void =>
-                      setFormState({
-                        ...formState,
-                        visible: !formState.visible,
-                      })
-                    }
-                  />
-                </Col>
-              </Row>
-              <Form.Label htmlFor="orgphoto">{t('displayImage')}</Form.Label>
-              <Form.Control
-                accept="image/*"
-                id="orgphoto"
-                className="mb-3"
-                name="photo"
-                type="file"
-                multiple={false}
-                onChange={async (e: React.ChangeEvent): Promise<void> => {
-                  const target = e.target as HTMLInputElement;
-                  const file = target.files && target.files[0];
-                  /* istanbul ignore else */
-                  if (file)
-                    setFormState({
-                      ...formState,
-                      image: await convertToBase64(file),
-                    });
-                }}
-                data-testid="organisationImage"
-              />
-              <Col className={styles.sampleOrgSection}>
-                <Button
-                  className={styles.orgCreationBtn}
-                  type="submit"
-                  value="invite"
-                  data-testid="submitOrganizationForm"
-                >
-                  {t('createOrganization')}
-                </Button>
-
-                <div className="position-relative">
-                  <hr />
-                  <span className={styles.orText}>{t('OR')}</span>
-                </div>
-                {userData &&
-                  ((userData.user.userType === 'ADMIN' &&
-                    userData.user.adminFor.length > 0) ||
-                    userData.user.userType === 'SUPERADMIN') && (
-                    <div className={styles.sampleOrgSection}>
-                      <Button
-                        className={styles.sampleOrgCreationBtn}
-                        onClick={() => triggerCreateSampleOrg()}
-                        data-testid="createSampleOrganizationBtn"
-                      >
-                        {t('createSampleOrganization')}
-                      </Button>
-                    </div>
-                  )}
-              </Col>
-            </Modal.Body>
-          </Form>
-        </Modal>{' '}
+        {/**
+         * Renders the `OrganizationModal` component.
+         *
+         * @param showModal - A boolean indicating whether the modal should be displayed.
+         * @param toggleModal - A function to toggle the visibility of the modal.
+         * @param formState - The state of the form in the organization modal.
+         * @param setFormState - A function to update the state of the form in the organization modal.
+         * @param createOrg - A function to handle the submission of the organization creation form.
+         * @param t - A translation function for localization.
+         * @param userData - Information about the current user.
+         * @param triggerCreateSampleOrg - A function to trigger the creation of a sample organization.
+         * @returns JSX element representing the `OrganizationModal`.
+         */}
+        <OrganizationModal
+          showModal={showModal}
+          toggleModal={toggleModal}
+          formState={formState}
+          setFormState={setFormState}
+          createOrg={createOrg}
+          t={t}
+          userData={userData}
+          triggerCreateSampleOrg={triggerCreateSampleOrg}
+        />
         {/* Plugin Notification Modal after Org is Created */}
         <Modal show={dialogModalisOpen} onHide={toggleDialogModal}>
           <Modal.Body>
